Changed vue lib to CLI, started work on components
