+++ package.json
@@ -1,0 +1,19 @@
+{
+  "name": "todolist-vue",
+  "version": "0.1.0",
+  "author": "Eugene Serb",
+  "license": "GNU GPL v3",
+  "private": true,
+  "scripts": {
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build"
+  },
+  "dependencies": {
+    "core-js": "^3.8.3",
+    "vue": "^3.2.13"
+  },
+  "devDependencies": {
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-service": "~5.0.0"
+  }
+}
