+++ src/App.vue
@@ -1,0 +1,19 @@
+<template>
+    <ToDoList />
+</template>
+
+<script>
+    import ToDoList from './components/ToDoList.vue';
+
+    export default {
+        name: 'App',
+        components: {
+            ToDoList: ToDoList,
+        },
+    };
+</script>
+
+<style>
+
+</style>
+
